+++ MeshLib/CMakeLists.txt
@@ -2,8 +2,9 @@
 GET_SOURCE_FILES(SOURCES_MESHLIB)
 GET_SOURCE_FILES(SOURCES_ELEMENTS Elements)
 GET_SOURCE_FILES(SOURCES_EDITING MeshEditing)
+GET_SOURCE_FILES(SOURCES_GENERATORS MeshGenerators)
 GET_SOURCE_FILES(SOURCES_QUALITY MeshQuality)
-SET ( SOURCES ${SOURCES_MESHLIB} ${SOURCES_ELEMENTS} ${SOURCES_EDITING} ${SOURCES_QUALITY})
+SET ( SOURCES ${SOURCES_MESHLIB} ${SOURCES_ELEMENTS} ${SOURCES_EDITING} ${SOURCES_GENERATORS} ${SOURCES_QUALITY})
 
 # Create the library
 ADD_LIBRARY(MeshLib STATIC ${SOURCES})
